requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=0.0.1
langchain-text-splitters>=0.0.1
langchain-community>=0.0.1
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pyarrow>=14.0.1
faiss-cpu>=1.7.4
pypdf>=3.0.0
